correcao do erro de id do codigo de tipo de pagamento

+++ database/migrations/2024_07_16_000320_create_abastecimentos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->integer('cupom');
             $table->integer('kmAnterior');
             $table->integer('kmAtual');
-            $table->integer('litros');
+            $table->double('litros');
             $table->double('valor');
             $table->string('pathFotoCupom')->nullable();
             $table->string('pathFotoHodometro')->nullable();
